camera: Makefile overhaul.
Upon trying to build only the OMX interface, it was discovered that some
necessary bits and pieces were only in the ALL interface. After attempting
to fix this, it became painfully obvious that the entire makefile should
needed a lot of sprucing up and updating.

The USB-only interface does not build, but it didn't before this either.
There's a number of (D)OMX ties in the common files that prevent it.
Considering nobody has noticed this, it's probably not really used.
Should somebody find they need it, they can try and fix it themselves.